+++ api/src/app.ts
@@ -21,7 +21,12 @@ app.use(express.urlencoded({ extended: true }));
 // app.use(cors());
 // app.use(cors({ credentials: true, origin: "http://localhost:5173/" }));
 // app.use(cors({ credentials: true, origin: "http://127.0.0.1:5173" }));
-app.use(cors({ origin: true, credentials: true }));
+app.use(
+  cors({
+    origin: true,
+    credentials: true,
+  })
+);
 
 import "./config/passport";