+++ api/src/interfaces/user.d.ts
@@ -3,7 +3,7 @@ import { Post } from "./post";
 import { Comment } from "./comment";
 
 export interface User {
-  id: UserId;
+  id?: UserId;
   role?: Role;
   username?: Username;
   email?: string;