+++ cms/src/app/App.tsx
@@ -1,7 +1,20 @@
 import React from "react";
-import { DataRouter, RouterProvider } from "react-router";
+import {
+  BrowserRouter,
+  DataRouter,
+  Route,
+  RouterProvider,
+  Routes,
+} from "react-router";
 
 import AuthProvider from "../hooks/useAuth";
+import DefaultLayout from "../layouts/DefaultLayout";
+import Dashboard from "../features/dashboard/Dashboard";
+
+import Home from "../features/home/Home";
+import About from "../features/about/About";
+import Faq from "../features/faq/Faq";
+import ProtectedLayout from "../layouts/ProtectedLayout";
 
 interface Props {
   router: DataRouter;
@@ -16,5 +29,23 @@ const App: React.FC<Props> = ({ router }) => {
     </div>
   );
 };
+
+/* const App: React.FC<Props> = () => {
+  return (
+    <div id="app">
+      <AuthProvider>
+        <Routes>
+          <Route index element={<Home />} />
+          <Route path="about" element={<About />} />
+
+          <Route element={<ProtectedLayout />}>
+            <Route path="login" element={<Dashboard />} />
+            <Route path="register" element={<Account />} />
+          </Route>
+        </Routes>
+      </AuthProvider>
+    </div>
+  );
+}; */
 
 export default App;